requirements.txt
pyrogram
tgcrypto
python-telegram-bot
httpx
python-dotenv
